refactor(bad-merge): there was a bad merge and mistake naming
