+++ src/app/core/provider/sentry.provider.spec.ts
@@ -1,10 +1,10 @@
 import { TestBed } from '@angular/core/testing';
 
-import { SENTRY_PROVIDER, SENTRY_TOKEN } from './sentry.factory';
+import { SENTRY_PROVIDER, SENTRY_TOKEN } from './sentry.provider';
 
 import * as Sentry from '@sentry/angular';
 
-describe('Factory: Sentry', () => {
+describe('Provider: Sentry', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
       providers: [SENTRY_PROVIDER],
