+++ src/app/core/provider/sentry.provider.ts
