+++ src/app/features/registry/pages/registration-recent-activity/registration-recent-activity.component.ts
@@ -9,7 +9,7 @@ import { DatePipe } from '@angular/common';
 import { ChangeDetectionStrategy, Component, computed, inject, OnDestroy, signal } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 
-import { ENVIRONMENT } from '@core/factory/environment.factory';
+import { ENVIRONMENT } from '@core/provider/environment.provider';
 import { CustomPaginatorComponent } from '@shared/components';
 import { ACTIVITY_LOGS_DEFAULT_PAGE_SIZE } from '@shared/constants/activity-logs';
 import {
