+++ src/testing/osf.testing.module.ts
@@ -8,7 +8,7 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NoopAnimationsModule, provideNoopAnimations } from '@angular/platform-browser/animations';
 import { provideRouter } from '@angular/router';
 
-import { WINDOW, windowFactory } from '@core/factory/window.factory';
+import { WINDOW, windowFactory } from '@core/provider/window.provider';
 
 import { DynamicDialogRefMock } from './mocks/dynamic-dialog-ref.mock';
 import { EnvironmentTokenMock } from './mocks/environment.token.mock';
